XC-349: split example into main.rs/lib.rs
